+++ src/main/java/com/moayong/api/domain/leaguemember/dto/response/MatchInfoResponse.java
@@ -10,7 +10,7 @@ public record MatchInfoResponse (
         Integer totalScore,
         Integer goalAmount,
         Integer rank,
-        Integer rate,
+        Float rate,
         Integer nextLevel,
         PromotionStatus promotionStatus
 ){
